attempting to add a combined l0prep/l1a workflow ... (read more) ...
I didn't want to deal with multiple outputs so the idea here is make the
next stage in the workflow end with a single output file L1A.  This will
take 12x L0 files as input before/target/after for apids 0/8/11/826.  It
will then run l0prep merging those together and then finally run the L1A
step. What this code adds is the ability to run these back to back.

ISSUE: previously we had the L1 step assumning it started from a
catalog, in order to make this work I hacked up the level1a-nocat script
such that it didn't need a catalog and then we can chain these
toghether. Is that a good idea? time will tell